+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added datatype `SKY_params_Distribution_AddressesDecoded`
 - Added datatype `SKY_params_Distribution_UnlockedAddressesDecoded`
 - Added datatype `SKY_params_Distribution_LockedAddressesDecoded`
+- Added datatype `SKY_params_Distribution_GetMainNetDistribution`
 
 ### Removed